Update Constraint properties to accept Float values
Interface Builder gives the impression that only integer values are allowed for the priority and constant properties. However, when a value is typed using a dot (e.g., 3.5) instead of a comma, the value is accepted.

Another issue is that Interface Builder may round the value to an integer visually, but this is only for display purposes — the underlying .storyboard file still stores the value as a float.

This commit updates the constant and priority properties to use Float, to better reflect:
- The actual representation in the storyboardThe behavior of the UIKit API, which also accepts floating-point values